@@ -13,13 +13,14 @@
     "color-thief-browser": "^2.0.2",
     "core-js": "^3.6.5",
     "echarts": "^4.9.0",
-    "element-ui": "^2.15.5",
+    "element-plus": "^1.1.0-beta.8",
     "moment": "^2.29.1",
     "monaco-editor": "^0.26.1",
     "register-service-worker": "^1.7.2",
-    "vue": "^2.6.12",
-    "vue-router": "^3.4.9",
-    "vuex": "^3.6.0"
+    "vue": "^3.2.8",
+    "@vue/compiler-sfc": "^3.2.8",
+    "vue-router": "^4.0.11",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.10.4",
@@ -42,8 +43,6 @@
     "less": "^3.12.2",
     "less-loader": "^6.2.0",
     "monaco-editor-webpack-plugin": "^4.1.1",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^9.0.3",
     "vue-cli-plugin-axios": "^0.0.4",
     "vue-cli-plugin-compression": "^1.1.4",
     "vue-cli-plugin-element": "^1.0.1",
