@@ -17,9 +17,11 @@
     </div>
     <div id="content">
       <div class="side"></div>
-      <transition>
-        <router-view />
-      </transition>
+      <router-view v-slot="{ Component }">
+        <transition>
+          <component :is="Component"></component>
+        </transition>
+      </router-view>
       <div class="side"></div>
     </div>
     <div class="about">
