@@ -33,9 +33,11 @@
         </div>
       </div>
       <div class="router">
-        <transition>
-          <router-view />
-        </transition>
+        <router-view v-slot="{ Component }">
+          <transition>
+            <component :is="Component"></component>
+          </transition>
+        </router-view>
       </div>
     </div>
   </div>
